% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_max}
\alias{min_max}
\title{Calculo del maximo y minimo}
\usage{
min_max(n, media, desvi)
}
\arguments{
\item{n}{tamaño de la muestra}

\item{media}{media de la poblacion}

\item{desvi}{desviacion de la poblacion}
}
\value{
un vector minimo y maximo
}
\description{
Calculo del maximo y minimo
}
\examples{
min_max(n=10,media=50,desvi=12)
min_max(n=100,media=3,desvi=8)

}
