% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_tres.R
\name{suma3}
\alias{suma3}
\alias{multi_tres}
\title{Esta funcion opera 3 valores}
\usage{
suma3(a, b, c)

multi_tres(a, b, c)
}
\arguments{
\item{a}{primero}

\item{b}{segundo}

\item{c}{tercero}
}
\value{
Esta funcion retorna el resultado
}
\description{
Esta funcion opera 3 valores
}
\examples{
suma3(a=1,b=9,c=7)
multi_tres(a=4,b=5,c=7)
}
